RAC-141: refactor registry should only be registry
